+++ compiler/rustc_codegen_llvm/src/gotoc/mir_to_goto/codegen/place.rs
@@ -5,9 +5,9 @@
 //! a place is an expression of specifying a location in memory, like a left value. check the cases
 //! in [codegen_place] below.
 
-use super::cbmc::goto_program::{Expr, Type};
 use super::typ::tuple_fld;
-use crate::gotoc::mir_to_goto::metadata::*;
+use crate::gotoc::cbmc::goto_program::{Expr, Type};
+use crate::gotoc::mir_to_goto::GotocCtx;
 use rustc_middle::{
     mir::{Field, Local, Place, ProjectionElem},
     ty::{self, Ty, TyS, VariantDef},
