+++ compiler/rustc_codegen_llvm/src/gotoc/mir_to_goto/mod.rs
@@ -5,3 +5,4 @@ pub mod monomorphize;
 //pub mod stubs;
 pub mod metadata;
 pub use metadata::GotocCtx;
+pub mod codegen;
